Refactored and moved files to appropriate locations
