+++ src/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import "./styles/global.css";
 import App from "./app/App.tsx";
-import { AlertDialogProvider } from "./components/dialog/AlertDialog/AlertDialogProvider.tsx";
+import { AlertDialogProvider } from "./components/AlertDialog/AlertDialogProvider.tsx";
 
 // For Google OAuth stretch goal 
 // import { GoogleOAuthProvider } from "@react-oauth/google";
